Move the vif device to its own package.
