+++ pkg/client/rootd/dns/resolved_linux.go
@@ -12,10 +12,10 @@ import (
 	"github.com/datawire/dlib/dlog"
 	"github.com/datawire/dlib/dtime"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/rootd/dbus"
-	"github.com/telepresenceio/telepresence/v2/pkg/vif"
+	"github.com/telepresenceio/telepresence/v2/pkg/vif/device"
 )
 
-func (s *Server) tryResolveD(c context.Context, dev vif.Device, configureDNS func(net.IP, *net.UDPAddr)) error {
+func (s *Server) tryResolveD(c context.Context, dev device.Device, configureDNS func(net.IP, *net.UDPAddr)) error {
 	// Connect to ResolveD via DBUS.
 	if !dbus.IsResolveDRunning(c) {
 		dlog.Error(c, "systemd-resolved is not running")
@@ -99,7 +99,7 @@ func (s *Server) tryResolveD(c context.Context, dev vif.Device, configureDNS fun
 	return g.Wait()
 }
 
-func (s *Server) updateLinkDomains(c context.Context, paths []string, dev vif.Device) error {
+func (s *Server) updateLinkDomains(c context.Context, paths []string, dev device.Device) error {
 	namespaces := make(map[string]struct{})
 	search := make([]string, 0)
 	for i, path := range paths {
